#45 - validator 사용성 개선
