--- src/main/java/com/sat/common/validation/StringLength.java
@@ -1,22 +1,0 @@
-package com.sat.common.validation;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = {StringLengthValidator.class})
-public @interface StringLength {
-
-    int min() default 0;
-    int max() default Integer.MAX_VALUE;
-    String message() default "길이가 올바르지 않습니다";
-    Class<?>[] groups() default { };
-    Class<? extends Payload>[] payload() default { };
-}
