+++ src/main/java/com/sat/study/application/dto/StudyGroupCreateRequest.java
@@ -1,6 +1,5 @@
 package com.sat.study.application.dto;
 
-import com.sat.common.validation.StringLength;
 import com.sat.study.domain.StudyCategory;
 import jakarta.validation.constraints.NotNull;
 
@@ -10,9 +9,8 @@ import java.time.LocalDateTime;
 import java.util.Set;
 
 public record StudyGroupCreateRequest(
-        @StringLength(min = 5, max = 20)
+        @StudyGroupTitle
         String title,
-        @StringLength(min = 10, max = 100)
         String contents,
         StudyCategory category,
         int maxCapacity,
