+++ src/main/java/com/sat/study/domain/StudyGroupTitle.java
@@ -1,6 +1,5 @@
 package com.sat.study.domain;
 
-import com.sat.common.validation.StringLengthValidator;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
@@ -11,14 +10,23 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Embeddable
 public class StudyGroupTitle {
-    private static final StringLengthValidator STRING_LENGTH_VALIDATOR =
-            new StringLengthValidator(5, 20);
+
+    public static final int MIN_LENGTH = 5;
+    public static final int MAX_LENGTH = 20;
+
+    private static final String EX_MESSAGE = String.format("스터디그룹 제목의 길이는 %d이상 %d이하이어야 합니다.", MIN_LENGTH, MAX_LENGTH);
 
     @Column(name = "title")
     private String value;
 
     public StudyGroupTitle(String value) {
-        STRING_LENGTH_VALIDATOR.validate(value, String.format("스터디그룹 제목의 길이는 %d이상 %d이하이어야 합니다. 입력값: %s", 5, 20, value));
+        if (isValid(value)) {
+            throw new IllegalArgumentException(EX_MESSAGE);
+        }
         this.value = value;
+    }
+
+    public static boolean isValid(String value) {
+        return value.length() < MIN_LENGTH || value.length() > MAX_LENGTH;
     }
 }
